requirements.txt
pandas==2.0.0
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
tiktoken==0.5.1
google-genai>=1.3.0
tqdm==4.66.1
python-dotenv==1.0.0 